Consolidate adapter config patcher into inference script
